+++ org.j3d-geom/src/main/java/org/j3d/geom/TorusGenerator.java
@@ -10,9 +10,11 @@
 package org.j3d.geom;
 
 // Standard imports
-import javax.vecmath.Vector3f;
+// None
 
 // Application specific imports
+
+import org.j3d.maths.vector.Vector3d;
 
 /**
  * A generator of Torus geometry values.
@@ -1061,7 +1063,7 @@ public class TorusGenerator extends GeometryGenerator
 
         int half = innerFacetCount / 2;
         float[] normals = data.normals;
-        Vector3f norm;
+        Vector3d norm;
         int count = 0;
         float[] origin = new float[3];
         int facet_offset = 0;
@@ -1080,54 +1082,54 @@ public class TorusGenerator extends GeometryGenerator
                                           count,
                                           oradiusCoordinates,
                                           facet_offset);
-                normals[count++] = norm.x;
-                normals[count++] = norm.y;
-                normals[count++] = norm.z;
+                normals[count++] = (float)norm.x;
+                normals[count++] = (float)norm.y;
+                normals[count++] = (float)norm.z;
 
                 norm = createRadialNormal(data.coordinates,
                                           count,
                                           oradiusCoordinates,
                                           facet_offset);
 
-                normals[count++] = norm.x;
-                normals[count++] = norm.y;
-                normals[count++] = norm.z;
+                normals[count++] = (float)norm.x;
+                normals[count++] = (float)norm.y;
+                normals[count++] = (float)norm.z;
 
                 norm = createRadialNormal(data.coordinates,
                                           count,
                                           oradiusCoordinates,
                                           facet_offset + 3);
 
-                normals[count++] = norm.x;
-                normals[count++] = norm.y;
-                normals[count++] = norm.z;
+                normals[count++] = (float)norm.x;
+                normals[count++] = (float)norm.y;
+                normals[count++] = (float)norm.z;
 
                 // triangle 2
                 norm = createRadialNormal(data.coordinates,
                                           count,
                                           oradiusCoordinates,
                                           facet_offset + 3);
-                normals[count++] = norm.x;
-                normals[count++] = norm.y;
-                normals[count++] = norm.z;
+                normals[count++] = (float)norm.x;
+                normals[count++] = (float)norm.y;
+                normals[count++] = (float)norm.z;
 
                 norm = createRadialNormal(data.coordinates,
                                           count,
                                           oradiusCoordinates,
                                           facet_offset + 3);
 
-                normals[count++] = norm.x;
-                normals[count++] = norm.y;
-                normals[count++] = norm.z;
+                normals[count++] = (float)norm.x;
+                normals[count++] = (float)norm.y;
+                normals[count++] = (float)norm.z;
 
                 norm = createRadialNormal(data.coordinates,
                                           count,
                                           oradiusCoordinates,
                                           facet_offset);
 
-                normals[count++] = norm.x;
-                normals[count++] = norm.y;
-                normals[count++] = norm.z;
+                normals[count++] = (float)norm.x;
+                normals[count++] = (float)norm.y;
+                normals[count++] = (float)norm.z;
 
                 facet_offset += 3;
             }
@@ -1137,54 +1139,54 @@ public class TorusGenerator extends GeometryGenerator
                                       count,
                                       oradiusCoordinates,
                                       facet_offset);
-            normals[count++] = norm.x;
-            normals[count++] = norm.y;
-            normals[count++] = norm.z;
+            normals[count++] = (float)norm.x;
+            normals[count++] = (float)norm.y;
+            normals[count++] = (float)norm.z;
 
             norm = createRadialNormal(data.coordinates,
                                       count,
                                       oradiusCoordinates,
                                       facet_offset);
 
-            normals[count++] = norm.x;
-            normals[count++] = norm.y;
-            normals[count++] = norm.z;
+            normals[count++] = (float)norm.x;
+            normals[count++] = (float)norm.y;
+            normals[count++] = (float)norm.z;
 
             norm = createRadialNormal(data.coordinates,
                                       count,
                                       oradiusCoordinates,
                                       0);
 
-            normals[count++] = norm.x;
-            normals[count++] = norm.y;
-            normals[count++] = norm.z;
+            normals[count++] = (float)norm.x;
+            normals[count++] = (float)norm.y;
+            normals[count++] = (float)norm.z;
 
             // triangle 2
             norm = createRadialNormal(data.coordinates,
                                       count,
                                       oradiusCoordinates,
                                       0);
-            normals[count++] = norm.x;
-            normals[count++] = norm.y;
-            normals[count++] = norm.z;
+            normals[count++] = (float)norm.x;
+            normals[count++] = (float)norm.y;
+            normals[count++] = (float)norm.z;
 
             norm = createRadialNormal(data.coordinates,
                                       count,
                                       oradiusCoordinates,
                                       0);
 
-            normals[count++] = norm.x;
-            normals[count++] = norm.y;
-            normals[count++] = norm.z;
+            normals[count++] = (float)norm.x;
+            normals[count++] = (float)norm.y;
+            normals[count++] = (float)norm.z;
 
             norm = createRadialNormal(data.coordinates,
                                       count,
                                       oradiusCoordinates,
                                       facet_offset);
 
-            normals[count++] = norm.x;
-            normals[count++] = norm.y;
-            normals[count++] = norm.z;
+            normals[count++] = (float)norm.x;
+            normals[count++] = (float)norm.y;
+            normals[count++] = (float)norm.z;
         }
 
         int tempVertCount = count;
@@ -1234,7 +1236,7 @@ public class TorusGenerator extends GeometryGenerator
 
         int half = innerFacetCount / 2;
         float[] normals = data.normals;
-        Vector3f norm;
+        Vector3d norm;
         int count = 0;
         float[] origin = new float[3];
         int facet_offset = 0;
@@ -1250,18 +1252,18 @@ public class TorusGenerator extends GeometryGenerator
                                       count,
                                       oradiusCoordinates,
                                       facet_offset);
-            normals[count++] = norm.x;
-            normals[count++] = norm.y;
-            normals[count++] = norm.z;
+            normals[count++] = (float)norm.x;
+            normals[count++] = (float)norm.y;
+            normals[count++] = (float)norm.z;
 
             norm = createRadialNormal(data.coordinates,
                                       count,
                                       oradiusCoordinates,
                                       facet_offset);
 
-            normals[count++] = norm.x;
-            normals[count++] = norm.y;
-            normals[count++] = norm.z;
+            normals[count++] = (float)norm.x;
+            normals[count++] = (float)norm.y;
+            normals[count++] = (float)norm.z;
 
             for(i = outerFacetCount; --i > 0; )
             {
@@ -1272,26 +1274,26 @@ public class TorusGenerator extends GeometryGenerator
                                           oradiusCoordinates,
                                           facet_offset);
 
-                normals[count++] = norm.x;
-                normals[count++] = norm.y;
-                normals[count++] = norm.z;
+                normals[count++] = (float)norm.x;
+                normals[count++] = (float)norm.y;
+                normals[count++] = (float)norm.z;
 
-                normals[count + 6] = norm.x;
-                normals[count + 7] = norm.y;
-                normals[count + 8] = norm.z;
+                normals[count + 6] = (float)norm.x;
+                normals[count + 7] = (float)norm.y;
+                normals[count + 8] = (float)norm.z;
 
                 norm = createRadialNormal(data.coordinates,
                                           count,
                                           oradiusCoordinates,
                                           facet_offset);
 
-                normals[count++] = norm.x;
-                normals[count++] = norm.y;
-                normals[count++] = norm.z;
+                normals[count++] = (float)norm.x;
+                normals[count++] = (float)norm.y;
+                normals[count++] = (float)norm.z;
 
-                normals[count++] = norm.x;
-                normals[count++] = norm.y;
-                normals[count++] = norm.z;
+                normals[count++] = (float)norm.x;
+                normals[count++] = (float)norm.y;
+                normals[count++] = (float)norm.z;
 
                 count += 3;
             }
@@ -1301,18 +1303,18 @@ public class TorusGenerator extends GeometryGenerator
                                       oradiusCoordinates,
                                       0);
 
-            normals[count++] = norm.x;
-            normals[count++] = norm.y;
-            normals[count++] = norm.z;
+            normals[count++] = (float)norm.x;
+            normals[count++] = (float)norm.y;
+            normals[count++] = (float)norm.z;
 
             norm = createRadialNormal(data.coordinates,
                                       count,
                                       oradiusCoordinates,
                                       0);
 
-            normals[count++] = norm.x;
-            normals[count++] = norm.y;
-            normals[count++] = norm.z;
+            normals[count++] = (float)norm.x;
+            normals[count++] = (float)norm.y;
+            normals[count++] = (float)norm.z;
         }
 
 
@@ -1356,7 +1358,7 @@ public class TorusGenerator extends GeometryGenerator
 
         int half = innerFacetCount / 2;
         float[] normals = data.normals;
-        Vector3f norm;
+        Vector3d norm;
         int count = 0;
         float[] origin = new float[3];
         int facet_offset = 0;
@@ -1375,18 +1377,18 @@ public class TorusGenerator extends GeometryGenerator
                                           oradiusCoordinates,
                                           facet_offset);
 
-                normals[count++] = norm.x;
-                normals[count++] = norm.y;
-                normals[count++] = norm.z;
+                normals[count++] = (float)norm.x;
+                normals[count++] = (float)norm.y;
+                normals[count++] = (float)norm.z;
 
                 norm = createRadialNormal(data.coordinates,
                                           count,
                                           oradiusCoordinates,
                                           facet_offset);
 
-                normals[count++] = norm.x;
-                normals[count++] = norm.y;
-                normals[count++] = norm.z;
+                normals[count++] = (float)norm.x;
+                normals[count++] = (float)norm.y;
+                normals[count++] = (float)norm.z;
 
                 facet_offset += 3;
             }
@@ -1396,18 +1398,18 @@ public class TorusGenerator extends GeometryGenerator
                                       oradiusCoordinates,
                                       0);
 
-            normals[count++] = norm.x;
-            normals[count++] = norm.y;
-            normals[count++] = norm.z;
+            normals[count++] = (float)norm.x;
+            normals[count++] = (float)norm.y;
+            normals[count++] = (float)norm.z;
 
             norm = createRadialNormal(data.coordinates,
                                       count,
                                       oradiusCoordinates,
                                       0);
 
-            normals[count++] = norm.x;
-            normals[count++] = norm.y;
-            normals[count++] = norm.z;
+            normals[count++] = (float)norm.x;
+            normals[count++] = (float)norm.y;
+            normals[count++] = (float)norm.z;
         }
 
         // Normals for the bottom half just copy the top half but have the
@@ -1444,7 +1446,7 @@ public class TorusGenerator extends GeometryGenerator
 
         int half = innerFacetCount / 2;
         float[] normals = data.normals;
-        Vector3f norm;
+        Vector3d norm;
         int count = 0;
         float[] origin = new float[3];
         int facet_count = 0;
@@ -1462,9 +1464,9 @@ public class TorusGenerator extends GeometryGenerator
                                           oradiusCoordinates,
                                           facet_count);
 
-                normals[count++] = norm.x;
-                normals[count++] = norm.y;
-                normals[count++] = norm.z;
+                normals[count++] = (float)norm.x;
+                normals[count++] = (float)norm.y;
+                normals[count++] = (float)norm.z;
 
                 facet_count += 3;
             }
